move insert inline back to fragment.js
